+++ lwjson/library.cmake
@@ -32,16 +32,16 @@ set(lwjson_include_DIRS
 )
 
 # Register core library to the system
-add_library(lwjson INTERFACE)
-target_sources(lwjson PUBLIC ${lwjson_core_SRCS})
-target_include_directories(lwjson INTERFACE ${lwjson_include_DIRS})
+add_library(lwjson)
+target_sources(lwjson PRIVATE ${lwjson_core_SRCS})
+target_include_directories(lwjson PUBLIC ${lwjson_include_DIRS})
 target_compile_options(lwjson PRIVATE ${LWJSON_COMPILE_OPTIONS})
 target_compile_definitions(lwjson PRIVATE ${LWJSON_COMPILE_DEFINITIONS})
 
 # Register lwjson debug module
-add_library(lwjson_debug INTERFACE)
-target_sources(lwjson_debug PUBLIC ${lwjson_debug_SRCS})
-target_include_directories(lwjson_debug INTERFACE ${lwjson_include_DIRS})
+add_library(lwjson_debug)
+target_sources(lwjson_debug PRIVATE ${lwjson_debug_SRCS})
+target_include_directories(lwjson_debug PUBLIC ${lwjson_include_DIRS})
 target_compile_options(lwjson_debug PRIVATE ${LWJSON_COMPILE_OPTIONS})
 target_compile_definitions(lwjson_debug PRIVATE ${LWJSON_COMPILE_DEFINITIONS})