+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Develop
 
+- Rework library CMake with removed INTERFACE type
+
 ## 1.7.0
 
 - Add clang-tidy